requirements.txt
requests
beautifulsoup4
PyGithub
pandas
streamlit
